bug fix del check_date

+++ includes/functions.inc.php
@@ -529,7 +529,7 @@ function gdrcd_check_time($time) {
 
     if ($time_hours == date('H')) {
         return date('i') - $time_minutes;
-    } elseif ($time_hours == (date('H') - 1) || $time_hours == (strftime('H') + 11)) {
+    } elseif ((int) $time_hours == (date('H') - 1) || $time_hours == ((int) strftime('H') + 11)) {
         return date('i') - $time_minutes + 60;
     }
